Unificacion de rutas en app.routes, funcional
