--- Back/src/routes/cliente.routes.js
@@ -1,10 +1,0 @@
-import express from "express";
-import { verificarToken } from "../middlewares/auth.js";
-import { getClientes, createCliente } from "../controllers/clienteController.js";
-
-const router = express.Router();
-
-router.get("/", verificarToken, getClientes);
-router.post("/", verificarToken, createCliente);
-
-export default router;
